Backend done

+++ backend/controller/goalController.js
@@ -1,10 +1,10 @@
 const Goal=require('../model/goalModel')
-
+const User=require('../model/userModel')
 // @desc Get goals
 // @route GET /api/goals
 // @access Private
 const getGoals = async(req,res)=>{
-    const goals=await Goal.find()
+    const goals=await Goal.find({user:req.user.id})
 
     res.status(200).json(goals)
 }
@@ -20,7 +20,8 @@ const postGoals = async(req,res)=>{
     }
 
     const goal=await Goal.create({
-        text:req.body.text
+        text:req.body.text,
+        user:req.user.id
     })
 
     res.status(200).json(goal)
@@ -31,9 +32,24 @@ const postGoals = async(req,res)=>{
 // @access Private
 const updateGoals = async(req,res)=>{
     const goal=await Goal.findById(req.params.id)
+
     if(!goal){
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    const user=await User.findById(req.user.id)
+
+    // Check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Check only logged in user matches goal user
+    if(goal.user.toString()!==user.id){
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     const updatedGoal=await Goal.findByIdAndUpdate(req.params.id,req.body,{new:true})
@@ -49,6 +65,21 @@ const deleteGoals = async(req,res)=>{
         res.status(400)
         throw new Error('Goal not found')
     }
+    
+    const user=await User.findById(req.user.id)
+
+    // Check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Check only logged in user matches goal user
+    if(goal.user.toString()!==user.id){
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
     await goal.remove()
     res.status(200).json({id: req.params.id})
     const deletedGoal=await Goal.findByIdAndDelete(req.params.id)
